Eliminated recordings.
I eliminated the code that referred to recordings. I will transfer this code to an independent project for clarity.

I also changed the subLSTM hidden_size parameter to an integer instead of a list with one value for each layer. I've done this for consistency with the LSTM implementation in PyTorch.